SCREAM: split nc-file-init into two files, and added capabilities
The script nc-file-init only inits an nc file, adding dimensions and
global attributes only. The new populate-nc-file script is now
responsible for adding vars to an nc file, by either:

 a. setting to a constant value
 b. evaluating a math expr (possibly involving other vars already stored)
 c. importing vars from another nc file with same resolution
 d. importing vars from another nc file with different resolution

Features b, c, and d make use of nco's executables. In particular,
d uses ncremap, which requires esmf.
