+++ components/scream/scripts/populate_nc_file.py
@@ -1,0 +1,373 @@
+from utils import expect, run_cmd_no_fail
+
+from netCDF4 import Dataset
+from nco import Nco
+
+import re, pathlib
+
+###############################################################################
+class PopulateNcFile(object):
+###############################################################################
+
+    ###########################################################################
+    def __init__(self,nc_file,import_file=None,map_file=None,overwrite=False,
+                 add_variables=None,import_variables=None,
+                 compute_variables=None,remove_variables=None):
+    ###########################################################################
+
+        self._overwrite = overwrite
+
+        self._ofile = pathlib.Path(nc_file).resolve().absolute()
+        self._ifile = pathlib.Path(import_file).resolve().absolute()
+        self._mfile = pathlib.Path(map_file).resolve().absolute()
+
+        self._avars     = add_variables
+        self._ivars     = import_variables
+        self._cvars     = compute_variables
+        self._rvars     = remove_variables
+
+        expect (self._ofile.exists(), "Error! File '{}' does not exist.".format(self._ofile))
+
+    ###########################################################################
+    def get_database(self,nc_file,mode):
+    ###########################################################################
+
+        expect (nc_file.exists(), "Error! Nc file {} does not exists.".format(nc_file))
+
+        ds = Dataset(nc_file,mode,persist=True)
+        expect ('ncol' in ds.dimensions,
+                "no ncol")
+        expect ('ncol' in ds.dimensions and ds.dimensions['ncol'].size>0,
+                "Error! NetCDF file '{}' does not contain a valid 'ncol' dimension."
+                .format(nc_file))
+        expect ('lev' in ds.dimensions and ds.dimensions['lev'].size>0,
+                "Error! NetCDF file '{}' does not contain a valid 'lev' dimension."
+                .format(nc_file))
+
+        return ds
+
+    ###########################################################################
+    def same_dims(self,dims1,dims2):
+    ###########################################################################
+        l = len(dims1)
+        if len(dims2)!=l:
+            return False
+
+        for i in range(1,l):
+            if dims1[i] != dims2[i]:
+                return False
+
+        return True
+
+    ###########################################################################
+    def check_var_name(self,var_name):
+    ###########################################################################
+        # re.match(r'^\w+$', string) is a very compact regexp check, to ensure
+        # the string only contains alphanumeric chars and underscores
+        expect(re.match(r'^\w+$', var_name),
+                "Error! Variable names must contain only alphanumeric characters or underscores.\n")
+
+    ###########################################################################
+    def check_overwrite_var(self,var_name):
+    ###########################################################################
+        ds = self.get_database(self._ofile,'r')
+        expect (not var_name in ds.variables.keys() or self._overwrite,
+                "Error! Variable '{}' already exists. To overwrite values, use -o flag.".format(var_name))
+        ds.close()
+
+    ###########################################################################
+    def add_variable(self,ds,name,dims,value):
+    ###########################################################################
+
+        # Check the var name is good (does not contain bad chars)
+        self.check_var_name(name)
+
+        # Check we are not overwriting (unless allowed)
+        self.check_overwrite_var(name)
+
+        if name in ds.variables.keys():
+            # If overwriting, make sure the dimensions are the same
+            var = ds.variables[name]
+            var_dims = var.get_dims()
+
+            expect (self.same_dims(var.get_dims(),dims),
+                    "Error! Trying to overwrite variable {} using wrong dimensions: ({}) insted of ({}).".
+                        format (name,
+                                ",".join(dims),
+                                ",".join([dim.name for dim in var_dims])))
+        else:
+            var = ds.createVariable(name,"f8",dims)
+        var[:] = value
+
+    ###########################################################################
+    def get_name(self,name_dims):
+    ###########################################################################
+        opn = name_dims.find('(')
+        cls = name_dims.find(')')
+
+        # Check format
+        expect (opn!=-1,"Error! Var declaration should be 'name(dim1,...,dimN)'.")
+        expect (cls!=-1,"Error! Var declaration should be 'name(dim1,...,dimN)'.")
+        expect (cls>opn,"Error! Var declaration should be 'name(dim1,...,dimN)'.")
+
+        name = name_dims[0:opn]
+        return name
+
+    ###########################################################################
+    def get_dims(self,name_dims):
+    ###########################################################################
+        opn = name_dims.find('(')
+        cls = name_dims.find(')')
+
+        # Check format
+        expect (opn!=-1,"Error! Var declaration should be 'name(dim1,...,dimN)'.")
+        expect (cls!=-1,"Error! Var declaration should be 'name(dim1,...,dimN)'.")
+        expect (cls>opn,"Error! Var declaration should be 'name(dim1,...,dimN)'.")
+        expect (cls==len(name_dims)-1,"Error! Var declaration should be 'name(dim1,...,dimN)'.")
+
+        dims = name_dims[opn+1:cls].split(',')
+        expect (len(dims)>0,"Error! Var declaration should be 'name(dim1,...,dimN)'.")
+
+        return dims
+
+    ###########################################################################
+    def is_vector_layout(self,dims):
+    ###########################################################################
+        valid = ["ncol", "lev", "ilev"]
+        for dim in dims:
+            if dim not in valid:
+                expect (dim.isdigit(), "Error! Unexpected dimension '{}'".format(dim))
+                return True
+        return False
+
+    ###########################################################################
+    def get_scalar_dims(self,dims):
+    ###########################################################################
+        valid = ["ncol", "lev", "ilev"]
+        s_dims = []
+        vec_dim_id = -1 
+        for i in range(0,len(dims)):
+            if dims[i] in valid:
+                s_dims.append(dims[i])
+            else:
+                expect (vec_dim_id==-1,
+                        "Error! Multiple integer extents found in dims specification '{}'.\n"
+                        "       Only vectors are supported, for non-scalar layouts.".format(dims))
+                vec_dim_id = i
+
+        expect(vec_dim_id>0, "Error! Something went wrong while detecting vector dim id from '{}'.".format(dims))
+
+        return vec_dim_id, s_dims
+
+    ###########################################################################
+    def get_values(self,vals_str,vec_dim):
+    ###########################################################################
+        if vals_str=="":
+            # User did not specify values. Use all zeros.
+            return [0]*vec_dim
+        else:
+            try:
+                # Try to convert vals_str to a single float. If successful,
+                # the user passed a single value for all vector components
+                value = float(vals_str)
+                return [value]*vec_dim
+            except ValueError:
+                # Didn't work. Then we must have a [v1,...,vN] format
+                opn = vals_str.find('[')
+                cls = vals_str.find(']')
+                expect (opn!=-1,"Error! Var values specification should be '..=val' or '..=[val1,...,valN]'.")
+                expect (cls!=-1,"Error! Var values specification should be '..=val' or '..=[val1,...,valN]'.")
+                expect (cls>opn,"Error! Var values specification should be '..=val' or '..=[val1,...,valN]'.")
+                expect (cls==len(vals_str)-1,"Error! Var values specification should be '..=val' or '..=[val1,...,valN]'.")
+                vals = vals_str[opn+1:cls].split(',')
+                expect (len(vals)==vec_dim,
+                        "Error! Var values specification has the wrong length: {} instead of {}."
+                        .format(len(vals),vec_dim))
+                try:
+                    values = [float(v) for v in vals]
+                    return values
+                except ValueError:
+                    expect(False, "Error! Something went wrong converting strings '{}' to floats.".format(vals))
+
+
+    ###########################################################################
+    def add_variables(self):
+    ###########################################################################
+
+        # Sync to file and close
+        ds = self.get_database(self._ofile,'a')
+        for item in self._avars:
+            if '=' in item:
+                # User provided initial values
+                name_dims_vals = item.split('=')
+                expect (len(name_dims_vals)==2, "Error! Invalid variable declaration: {}".format(item))
+                name_dims = name_dims_vals[0]
+                vals_str = name_dims_vals[1]
+            else:
+                name_dims = item
+                vals_str = ""
+
+            # From the string name(dim1,...,dimN) extract name and dimensions
+            name = self.get_name(name_dims)
+            dims = self.get_dims(name_dims)
+
+            is_vector = self.is_vector_layout(dims)
+
+
+            if is_vector:
+                # From the list (dim1,...,dimN), check if it is a vector field,
+                # and if so, get the idx of the vector dimension, the extent
+                # along that dimension, and the dims list without the vector dim.
+                vec_dim_id, scalar_dims = self.get_scalar_dims(dims)
+
+                vec_dim = 1 if vec_dim_id==-1 else int(dims[vec_dim_id])
+
+                # From the string after the = (if any), get the initialization
+                # values. The string can be a single value (for scalar or vector
+                # fields) or a list of values [v1,...,vn] (for vector field)
+                values = self.get_values(vals_str,vec_dim)
+
+                for i in range(0,len(values)):
+                    self.add_variable(ds,"{}_{}".format(name,i),scalar_dims,values[i])
+            else:
+                value = 0.0 if vals_str=="" else float(vals_str)
+                self.add_variable(ds,name,dims,value)
+
+        ds.sync()
+        ds.close()
+
+    ###########################################################################
+    def check_dims(self,ds_in,ds_out,dims):
+    ###########################################################################
+        for dim in dims:
+            expect (dim in ds_out.dimensions,
+                    "Error! Dimension {} not found in the output file '{}'.".format(dim,self._ofile))
+            expect (ds_in.dimensions[dim].size==ds_out.dimensions[dim].size,
+                    "Error! Dimension {} in input file '{}' has a different extent than in output file '{}'.\n"
+                    "   - {}: {}\n"
+                    "   - {}: {}".format(dim,self._ifile,self._ofile,
+                                         self._ifile,ds_in.dimensions[dim].size,
+                                         self._ofile,ds_out.dimensions[dim].size))
+
+    ###########################################################################
+    def import_variables(self):
+    ###########################################################################
+
+        if len(self._ivars)>0:
+            expect (self._ifile.exists(),
+                    "Error! Import file '{}' does not exist.".format(self._ifile))
+
+            ds_out = self.get_database(self._ofile,'a')
+            ds_in  = self.get_database(self._ifile,'r')
+
+            expect ('ncol' in ds_in.dimensions,
+                    "Error! 'ncol' not found in input file dimensions'")
+            expect ('lev' in ds_in.dimensions,
+                    "Error! 'lev' not found in input file dimensions'")
+
+            ncol_out = ds_out.dimensions['ncol'].size
+            nlev_out = ds_out.dimensions['lev'].size
+            ncol_in  = ds_in.dimensions['ncol'].size
+            nlev_in  = ds_in.dimensions['lev'].size
+
+            ds_in.close()
+            ds_out.close()
+
+            expect (nlev_in==nlev_out,
+                    "Error! Vertical remapping unavailable, due to ncremap assumption that level idx strides slower than column idx.")
+
+            if ncol_in==ncol_out:
+                self.import_variables_no_remap(self._ifile)
+            else:
+                self.import_variables_horiz_remap()
+
+
+    ###########################################################################
+    def import_variables_no_remap(self,ifile):
+    ###########################################################################
+
+        ds_out = self.get_database(self._ofile,'a')
+        ds_in  = self.get_database(ifile,'r')
+        for item in self._ivars:
+            if '=' in item:
+                tokens = item.split('=')
+                expect(len(tokens)==2,
+                       "Error! Import variable with either 'name' or 'name=name_in', where name_in is\n"
+                       "       is the var name in the input file, and name is the var name in the output file.")
+                name_out = tokens[0]
+                name_in  = tokens[1]
+            else:
+                name_out = item
+                name_in  = item
+
+            var = ds_in.variables[name_in]
+
+            # Make sure this var's dims are in our output nc file
+            self.check_dims(ds_in,ds_out,var.dimensions)
+            self.check_overwrite_var(name_out)
+            if name_out not in ds_out.variables.keys():
+                self.check_var_name(name_out)
+                ds_out.createVariable(name_out,var.dtype,var.dimensions)
+
+            ds_out.variables[name_out][:] = var[:]
+
+        ds_in.close()
+        ds_out.sync()
+        ds_out.close()
+
+    ###########################################################################
+    def import_variables_horiz_remap(self):
+    ###########################################################################
+        
+        # Use a temp file, cause ncremap adds a lot of auxiliary fields
+        temp_file = pathlib.Path("./pncf_tmp.nc").resolve().absolute()
+        cmd = " ncremap -m {} -i {} -o {} -v {}".\
+                format(self._mfile,self._ifile,temp_file,",".join(self._ivars))
+        run_cmd_no_fail(cmd)
+
+        # Import only desired vars from temp file to the actual nc file
+        self.import_variables_no_remap(temp_file)
+
+        # Clean up temp file
+        run_cmd_no_fail("rm {}".format(temp_file))
+
+    ###########################################################################
+    def compute_variables(self):
+    ###########################################################################
+
+        for expr in self._cvars:
+            # Split the expression, to get the output var name
+            tokens = expr.split('=')
+            expect(len(tokens)==2,"Error! Compute variables with 'var=expr' syntax.")
+
+            var_name = tokens[0]
+
+            self.check_var_name(var_name)
+            self.check_overwrite_var(var_name)
+
+            Nco().ncap2(str(self._ofile),output=str(self._ofile),spt=expr)
+
+    ###########################################################################
+    def remove_variables(self):
+    ###########################################################################
+        
+        if len(self._rvars)>0:
+            Nco().ncks(str(self._ofile),output=str(self._ofile),exclude=True,force=True,variable=",".join(self._rvars))
+
+    ###########################################################################
+    def run(self):
+    ###########################################################################
+
+        # Add vars, initing to constant value
+        self.add_variables()
+
+        # Import vars from another file
+        self.import_variables()
+
+        # Compute variables from math expressions (possibly involving other stored vars)
+        self.compute_variables()
+
+        # Remove variables
+        self.remove_variables()
+
+        return True
